Moving all public classes outside the `exporters` package
