+++ disk-buffering/src/main/java/io/opentelemetry/contrib/disk/buffering/StoredBatchExporter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.disk.buffering.exporters;
+package io.opentelemetry.contrib.disk.buffering;
 
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
