+++ disk-buffering/src/test/java/io/opentelemetry/contrib/disk/buffering/SpanDiskExporterTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.disk.buffering.exporters;
+package io.opentelemetry.contrib.disk.buffering;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
@@ -11,25 +11,25 @@ import static org.mockito.Mockito.mock;
 import io.opentelemetry.contrib.disk.buffering.internal.storage.TestData;
 import io.opentelemetry.contrib.disk.buffering.storage.StorageConfiguration;
 import io.opentelemetry.sdk.common.CompletableResultCode;
-import io.opentelemetry.sdk.logs.export.LogRecordExporter;
+import io.opentelemetry.sdk.trace.export.SpanExporter;
 import java.io.File;
 import java.io.IOException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-class LogRecordDiskExporterTest {
-  private LogRecordExporter wrapped;
-  private LogRecordDiskExporter exporter;
+class SpanDiskExporterTest {
+  private SpanExporter wrapped;
+  private SpanDiskExporter exporter;
   private static final StorageConfiguration STORAGE_CONFIGURATION =
       TestData.getDefaultConfiguration();
-  private static final String STORAGE_FOLDER_NAME = "logs";
+  private static final String STORAGE_FOLDER_NAME = "spans";
   @TempDir File rootDir;
 
   @BeforeEach
   public void setUp() throws IOException {
     wrapped = mock();
-    exporter = LogRecordDiskExporter.create(wrapped, rootDir, STORAGE_CONFIGURATION);
+    exporter = SpanDiskExporter.create(wrapped, rootDir, STORAGE_CONFIGURATION);
   }
 
   @Test
